requirements.txt
Flask
flask-mysql-connector
mysql-connector-python
python-dotenv
Werkzeug
Flask-JWT-Extended